move Task to its own module
